feat(project): 添加项目点赞和收藏数量显示
- 在 ProjectDetail.vue 中添加点赞和收藏数量的显示
- 在后端 ProjectFacade 中增加获取点赞和收藏数量的方法- 在 ProjectLikeService 和 ProjectFavoriteService 中添加按项目 ID 统计数量的方法
- 更新 ProjectsDetailResp 中点赞和收藏数量的类型为 Long

+++ my-project-backend/src/main/java/com/example/entity/resp/ProjectsDetailResp.java
@@ -30,6 +30,6 @@ public class ProjectsDetailResp {
     private String creatorName;
     private Boolean myLike;
     private Boolean myFavorite;
-    private Integer likeCount;
-    private Integer favoriteCount;
+    private Long likeCount;
+    private Long favoriteCount;
 }
